% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{bccamtrap_app}
\alias{bccamtrap_app}
\title{Launch Shiny app}
\usage{
bccamtrap_app()
}
\description{
Interactive interface for Data import, QA, and esport.
}
\examples{
\dontrun{bccamtrap_app()}

}
\concept{shiny_app}
