% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{bccamtrapp}
\alias{bccamtrapp}
\title{Launch Shiny app}
\usage{
bccamtrapp(...)
}
\arguments{
\item{...}{passed on to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\description{
Interactive interface for Data import, QA, and export.
}
\examples{
\dontrun{bccamtrapp()}

}
\concept{shiny_app}
