% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-deployment-images.R
\name{check_deployment_images}
\alias{check_deployment_images}
\title{Check for mismatched deployment labels in session data and image data}
\usage{
check_deployment_images(sessions, image_data)
}
\arguments{
\item{sessions}{sessions data, as created by \code{\link[=make_sample_sessions]{make_sample_sessions()}}}

\item{image_data}{data.frame of image sequence data, as read via \code{\link[=read_image_data]{read_image_data()}}}
}
\value{
a list of two vectors containing mismatched deployment labels
}
\description{
This function is mainly called for its messaging - alerting you to deployment labels that are
in the sample session data but not in the image data, and vice versa.
}
