% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-spatial.R
\name{check_stations_spatial}
\alias{check_stations_spatial}
\title{Check spatial attributes of Sample Station information}
\usage{
check_stations_spatial(stations, quant_thresh = 0.9, dist_thresh = NULL)
}
\arguments{
\item{stations}{an \code{sf} data.frame as read in by \code{\link[=read_sample_station_info]{read_sample_station_info()}}}

\item{quant_thresh}{The quantile above which the distance of a station from
the other stations (compared to the average distance between stations)
will be flagged as an outlier. Default \code{0.85}.}

\item{dist_thresh}{The distance (in m) above which the distance of a station
from the other stations will be flagged as an outlier. By default the
quantile is used, but can be overriden by setting this to a numeric value
in metres.}
}
\value{
the input data.frame, with a new \code{logical} column \code{spatial_outlier} appended.
}
\description{
This checks for inconsistent placement of stations to flag potentially
incorrect location information
}
\seealso{
\code{\link[=map_stations]{map_stations()}}
}
