% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_migration_episodes}
\alias{collapse_migration_episodes}
\title{Collapse motion detection events into episodes}
\usage{
collapse_migration_episodes(df)
}
\arguments{
\item{df}{a data frame of image data, read in with \code{read_image_data()}.}
}
\value{
a data frame with motion detection events collapsed into 1 row per
episode, and summary information about each episode (e.g. duration, number
of triggers, total count).
}
\description{
Collapse motion detection events into episodes
}
