% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_wallow_episodes}
\alias{collapse_wallow_episodes}
\title{Collapse Wallow Episodes}
\usage{
collapse_wallow_episodes(df)
}
\arguments{
\item{df}{a data frame of wallow image data, read in with \code{read_image_data()}.}
}
\value{
a data frame with wallow image data collapsed into 1 row per
episode, and summary information about each episode (e.g. duration, number
of triggers, total count, social and wallow behaviour, ...).
}
\description{
Collapse Wallow Episodes
}
