% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{fill_spi_template}
\alias{fill_spi_template}
\title{Write all data to a new SPI submission file}
\usage{
fill_spi_template(
  sample_station_info,
  camera_info,
  camera_setup_checks,
  image_data,
  file,
  template = system.file("GENERIC_wildlife_camera_template_v2021.xlsm", package =
    "bccamtrap")
)
}
\arguments{
\item{sample_station_info}{object of class \code{sample_station_info}}

\item{camera_info}{object of class \code{camera_info}}

\item{camera_setup_checks}{object of class \code{camera_setup_checks}}

\item{image_data}{object of class \code{image_data}}

\item{file}{path to the output csv file}
}
\value{
path to the output \code{file}
}
\description{
This will write only the default columns for each of the tabs in the SPI
template. For more fine-grained control, use \code{\link[=write_to_spi_sheet]{write_to_spi_sheet()}} on each
of \code{sample_station_info}, \code{camera_info}, \code{cam_setup_checks}, \code{image_data}.
}
