% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{fill_spi_template_ff}
\alias{fill_spi_template_ff}
\title{Write data to a new SPI submission file from field form data}
\usage{
fill_spi_template_ff(
  sample_station_info,
  deployments,
  image_data = NULL,
  wlrs_project_name = NULL,
  file,
  template = default_spi_template()
)
}
\arguments{
\item{sample_station_info}{object of class \code{c("sample_station_info", "field-form")}}

\item{deployments}{object of class \code{c("deployments", "field-form")}}

\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{wlrs_project_name}{the name of the project to write, if there is more
than one project in \code{sample_station_info} and/or \code{deployments}.}

\item{file}{path to the output file (\verb{.xls*})}

\item{template}{SPI submission template to use. Default is included in the
package, accessed by
\code{system.file("GENERIC_wildlife_camera_template_v2021.xlsm", package = "bccamtrap")}}
}
\value{
path to the output \code{file}
}
\description{
This will write only the default columns for each of the tabs in the SPI
template.
}
\details{
\code{sample_station_info} and \code{deployments} are required. If you want to only
write to the metadata tabs and not the Sequence Image Data, you can leave the
\code{image_data} argument as \code{NULL}, and write to the file another time with
\code{\link[=write_to_spi_sheet]{write_to_spi_sheet()}}.
}
