% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{make_deployments}
\alias{make_deployments}
\title{Create a spatial data frame of deployments}
\usage{
make_deployments(path, as_sf = TRUE)
}
\arguments{
\item{path}{path to the RISC workbook}

\item{as_sf}{should the data be returned as an \code{sf} object of the station locations? Default \code{TRUE}}
}
\value{
a spatial data.frame (\code{sf}) of deployments
}
\description{
Merge data from "Sample Station Info" and "Camera Setup and Checks" tabs to
create deployments - the time that a camera was deployed and running
at a site. "Invalid" deployments are flagged when there is no sampling end date
in "Camera Setup and Checks" or the sampling period is 0 days
}
