% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{make_sample_sessions}
\alias{make_sample_sessions}
\title{Identify distinct sample sessions from timelapse image data}
\usage{
make_sample_sessions(
  image_data,
  sample_start_date = NULL,
  sample_end_date = NULL
)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{sample_start_date}{a custom start date. Note that this will apply the
same start date to all deployments/sessions in the data.}

\item{sample_end_date}{a custom end date. Note that this will apply the same
start date to all deployments/sessions in the data.}
}
\value{
a data.frame of class \code{sample_sessions} with one row (sample session) per deployment
}
\description{
For each deployment label, this function:
\itemize{
\item Sets sampling_start as first image date
\item Counts photos (total, and motion-detection)
\item Determines gaps in sampling period due to obscured lens
\item Determines total length of sample period (number of days with Time Lapse photos - number of days with lens obscured)
}
}
