% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessions.R
\name{make_sessions}
\alias{make_sessions}
\title{Identify distinct sample sessions from deployments}
\usage{
make_sessions(deployments, image_data, as_sf = TRUE)
}
\arguments{
\item{deployments}{deployments data, as created by \code{\link[=make_deployments]{make_deployments()}}}

\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{as_sf}{should the data be returned as an \code{sf} object of the station locations? Default \code{TRUE}}
}
\value{
a data.frame of class \code{sample_sessions} with one row (sample session) per deployment
}
\description{
This function:
\itemize{
\item Sets sampling_start as deployment_start
\item Notes dates of first and last photos of deployment
\item Counts photos (total, and motion-detection)
\item Determines if the sampling period is less than the deployment period
\item Determines gaps in sampling period due to obscured lens
\item Determines total length of sample period (last photo date - first photo date - number of days with lens obscured)
}
}
