% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-spatial.R
\name{map_stations}
\alias{map_stations}
\title{Visualize stations on an interactive map}
\usage{
map_stations(stations)
}
\arguments{
\item{stations}{an \code{sf} data.frame as read in by \code{\link[=read_sample_station_info]{read_sample_station_info()}}}
}
\value{
a mapview object
}
\description{
If the input data.frame has been run through \code{\link[=qa_stations_spatial]{qa_stations_spatial()}},
potential spatial outliers will be visually flagged.
}
\seealso{
\code{\link[=qa_stations_spatial]{qa_stations_spatial()}}
}
