% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-deployment-images.R
\name{plot_deployment_detections}
\alias{plot_deployment_detections}
\title{Plot image timestamps over deployment periods}
\usage{
plot_deployment_detections(
  deployments,
  image_data,
  date_breaks = "1 month",
  interactive = FALSE
)
}
\arguments{
\item{deployments}{deployments data, as created by \code{\link[=make_deployments]{make_deployments()}}}

\item{image_data}{data.frame of image sequence data, as read via \code{\link[=read_image_data]{read_image_data()}}}

\item{date_breaks}{How to break up the dates on the x axis. See \code{\link[scales:date_breaks]{scales::date_breaks()}}. Default \code{"1 month"}}

\item{interactive}{should the plot be interactive? Default \code{FALSE}}
}
\value{
a \code{ggplot2} object if \code{interactive = FALSE}, a \code{plotly} object if \code{TRUE}
}
\description{
Plot detections over sample deployments to check for misaligned time stamps
}
