% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-deployment-images.R
\name{plot_diel_activity}
\alias{plot_diel_activity}
\title{Plot daily activity of species}
\usage{
plot_diel_activity(image_data, interactive = FALSE)
}
\arguments{
\item{image_data}{data.frame of image sequence data, as read via \code{\link[=read_image_data]{read_image_data()}}}

\item{interactive}{should the plot be interactive? Default \code{FALSE}}
}
\value{
a \code{ggplot2} object if \code{interactive = FALSE}, a \code{plotly} object if \code{TRUE}
}
\description{
Plot daily activity of species
}
