% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-images.R
\name{plot_snow}
\alias{plot_snow}
\title{Plot snow depths across deployments}
\usage{
plot_snow(image_data, date_breaks = "1 month", interactive = FALSE)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{date_breaks}{How to break up the dates on the x axis. See \code{\link[scales:date_breaks]{scales::date_breaks()}}. Default \code{"1 month"}}

\item{interactive}{should the plot be interactive? Default \code{FALSE}}
}
\value{
a \code{ggplot2} object if \code{interactive = FALSE}, a \code{plotly} object if \code{TRUE}
}
\description{
Plot snow depths across deployments
}
