% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-deployment-images.R
\name{qa_deployment_images}
\alias{qa_deployment_images}
\title{Check for mismatched deployment labels in session data and image data}
\usage{
qa_deployment_images(deployments, image_data)
}
\arguments{
\item{deployments}{deployments data, as created by \code{\link[=make_deployments]{make_deployments()}}}

\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}
}
\value{
a list of two vectors containing mismatched deployment labels
}
\description{
This function is mainly called for its messaging - alerting you to deployment labels that are
in the sample session data but not in the image data, and vice versa.
}
