% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-images.R
\name{qa_image_data}
\alias{qa_image_data}
\title{Perform a series of QA checks on the Timelapse Image data}
\usage{
qa_image_data(image_data, exclude_human_use = TRUE, check_snow = TRUE)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{exclude_human_use}{should images where \code{human_use_type} indicates a motion
detection event from humans be excluded from the count/species checks? Default \code{TRUE}.}

\item{check_snow}{show \code{QA} be performed on snow measurements? Default \code{TRUE}}
}
\value{
input data frame with additional \verb{QA_*} columns appended, and
subset only to rows where a QA issue was flagged.
}
\description{
\itemize{
\item Check for blanks in key fields (study area, station label, deployment date,
surveyor, trigger mode, temperature, episode)
\item Species detected with no count data
\item Count data with no species
\item Sum of subcount fields equals Total Count
\item Multiple entries under same Episode number (indicating possible double entry)
\item Ensure dates for timelapse images are continuous and in order.
\item Snow data
\itemize{
\item No blanks unless lens obscured is TRUE
\item Look for snow depth outliers (e.g., 10, 10, 110, 10, 15, 20)
}
}
}
