% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rai.R
\name{rai_by_time}
\alias{rai_by_time}
\title{Calculate RAI over a window of time, optionally rolling}
\usage{
rai_by_time(
  image_data,
  by = c("date", "week", "month", "year"),
  roll = FALSE,
  k = 7,
  by_species = TRUE,
  species = NULL,
  by_deployment = FALSE,
  deployment_label = NULL,
  sample_start_date = NULL,
  sample_end_date = NULL
)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{by}{time to aggregate by: One of \code{"date"} (default), \code{"week"}, \code{"month"}, or \code{"year"}.}

\item{roll}{should it use a rolling window? Default \code{FALSE}}

\item{k}{the size of the rolling window. Default \code{7}.}

\item{by_species}{Should an RAI be calculated for each species (\code{TRUE},
default), or one RAI for all species (\code{FALSE})}

\item{species}{Optionally, or more species as a character vector. Default
\code{NULL} to calculate RAI for each species detected.}

\item{by_deployment}{Should it be calculated by \code{deployment}. Default \code{FALSE}}

\item{deployment_label}{Optionally, one or more deployment labels to select.
Default \code{NULL} to use all deployments.}

\item{sample_start_date}{a custom start date. Note that this will apply the
same start date to all deployments/sessions in the data.}

\item{sample_end_date}{a custom end date. Note that this will apply the same
start date to all deployments/sessions in the data.}
}
\value{
a data.frame of above calculated metrics
}
\description{
This function takes a data frame of image data and calculates
snow, temperature, effort, count, and RAI metrics aggregated
over a specified time window. These can optionally be calculated
using a moving window ("rolling") calculation.
}
