% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-metadata.R
\name{read_cam_setup_checks}
\alias{read_cam_setup_checks}
\title{Read the "Camera Setup and Checks" tab from the RISC worksheet.}
\usage{
read_cam_setup_checks(path, ...)
}
\arguments{
\item{path}{path to the RISC workbook}

\item{...}{arguments passed on to \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\value{
a \code{data.frame} of camera setup information
}
\description{
This will read in the camera setup information from a RISC worksheet following the
'v20230518' template
}
