% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-data.R
\name{read_image_data}
\alias{read_image_data}
\title{Read image data from a collection of csvs from TimeLapse}
\usage{
read_image_data(path, pattern, recursive = FALSE, template = NULL, ...)
}
\arguments{
\item{path}{path to directory of image files, a single .csv file, or a character
vector of .csv files.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will read. Default \code{FALSE}.}

\item{recursive}{should files found within subfolders of \code{path} also be read?}

\item{template}{path to a "\code{.tdb}" TimeLapse Template file. Optional; if not provided,
the function will attempt to identify the appropriate internal template
based on the file names in \code{path}.}

\item{...}{arguments passed on to \code{\link[readr:read_delim]{readr::read_csv()}}}
}
\value{
a \code{data.frame} of Timelapse image data from the files found in \code{path}.
The data.frame will have an "image_data" class, and an attribute "template"
with the name of the template used to read the data.
}
\description{
This function reads in image data from a directory of csv files, or a
character vector of csv file paths. It attempts to identify the appropriate
TimeLapse template to use based on the file names, but a specific template
can also be provided. The function does some basic standardization and
cleaning of the data, and returns a data.frame with an "image_data" class
and an attribute "template" with the name of the template used to read the data.
}
\details{
In addition to reading in the data, this function also:
\itemize{
\item does basic standardization of trigger mode values,
\item creates numeric snow depth columns,
\item checks for the presence of a \code{total_count_episode} column, adding one if missing,
\item standardizes column names to snake_case.
\item If the data has separate \code{Date} and \code{Time} columns instead of a combined \code{DateTime} column,
these will be reconciled into a single \code{DateTime} column.
}

For wallow data, this also removes motion-activated static images,
and only keeps the motion-activated video records and the time-lapse static image records.
}
