% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-metadata.R
\name{read_project_info}
\alias{read_project_info}
\title{Read project-level data from the RISC worksheet 'Project Information' tab}
\usage{
read_project_info(path, ...)
}
\arguments{
\item{path}{path to the RISC workbook}

\item{...}{arguments passed on to \code{\link[readxl:read_excel]{readxl::read_excel()}}}
}
\value{
a \code{data.frame} of project-level information
}
\description{
This will read in the project information from a RISC worksheet following the
'v20230518' template
}
