% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-field-forms.R
\name{read_sample_station_csv}
\alias{read_sample_station_csv}
\title{Read csv output from sample station field form}
\usage{
read_sample_station_csv(path, wlrs_project_name = NULL, as_sf = TRUE)
}
\arguments{
\item{path}{path to the csv file exported from the field form}

\item{wlrs_project_name}{If you want to subset to a particular project or projects,
supply the WLRS Project Names(s) as a character vector. If \code{NULL} (the default),
reads all rows in the csv.}

\item{as_sf}{should the data be returned as an \code{sf} object of the station locations? Default \code{TRUE}}
}
\value{
a data.frame (sf object if \code{as_sf = TRUE}) containing sample station csv contents
}
\description{
Read csv output from sample station field form
}
