% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rai.R
\name{sample_rai}
\alias{sample_rai}
\title{Calculate Relative Activity Index (RAI)}
\usage{
sample_rai(
  image_data,
  deployment_label = NULL,
  species = NULL,
  by_deployment = TRUE,
  by_species = TRUE,
  sample_start_date = NULL,
  sample_end_date = NULL
)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{deployment_label}{One or more deployment labels to select}

\item{species}{One or more species as a character vector. Default \code{NULL} to
calculate RAI for each species detected.}

\item{by_deployment}{Should an RAI be calculated for each deployment
label (\code{TRUE}, default), or one RAI for all deployment labels (\code{FALSE})}

\item{by_species}{Should an RAI be calculated for each species (\code{TRUE},
default), or one RAI for all species (\code{FALSE})}

\item{sample_start_date}{a custom start date. Note that this will apply the
same start date to all deployments/sessions in the data.}

\item{sample_end_date}{a custom end date. Note that this will apply the same
start date to all deployments/sessions in the data.}
}
\value{
a data.frame of RAI, by deployment label and species (if \code{by_species = TRUE})
}
\description{
Calculate the RAI for sample sessions and species. Optionally set start date
and end date, deployment label(s), and species. You can choose to calculate
RAI by species and deployment label, or as a whole.
}
