% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-metadata.R
\name{to_sf}
\alias{to_sf}
\title{Convert Sample Station Info or Camera Info to a spatial \code{sf} boject
#'}
\usage{
to_sf(x, ...)
}
\arguments{
\item{x}{Object of class \code{sample_station_info} or \code{camera_info}}

\item{...}{arguments passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}}
}
\value{
input as a an \code{sf} object
}
\description{
Convert Sample Station Info or Camera Info to a spatial \code{sf} boject
#'
}
