% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_image_data}
\alias{write_image_data}
\title{Write image data to a csv file}
\usage{
write_image_data(image_data, file, na = "", ...)
}
\arguments{
\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{file}{path to the output csv file}

\item{na}{How should missing values be written. Default empty (\code{""})}

\item{...}{other parameters passed on to \code{\link[readr:write_delim]{readr::write_csv()}}}
}
\value{
input \code{image_data}, invisibly
}
\description{
This is a very thin wrapper around \code{\link[readr:write_delim]{readr::write_csv()}}. You can use that
function directly for more control
}
\examples{
\dontrun{
write_image_data(image_data, "my_spi_submission.xlsx")
write_image_data(image_data, "my_spi_submission.xlsx", Surveyor = surveyor,
   `Survey Observation Photos` = file)
}
}
