% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_to_spi_sheet}
\alias{write_to_spi_sheet}
\title{Write a single data frame to SPI submission template}
\usage{
write_to_spi_sheet(x, file, ..., template = default_spi_template())
}
\arguments{
\item{x}{object of class \code{sample_station_info}, \code{camera_info}, \code{cam_setup_checks}, or \code{image_data}}

\item{file}{path to the output file (\verb{.xls*})}

\item{...}{extra columns in \code{x} to write out. Must be paired column
names in the form \code{`Destination Column` = data_column}. If the left-hand side
is a syntactically valid name it can be provided as-is, but if it has spaces in it
it must be wrapped in backticks or quotes. See examples.}

\item{template}{SPI submission template to use. Default is included in the
package, accessed by
\code{system.file("GENERIC_wildlife_camera_template_v2021.xlsm", package = "bccamtrap")}}
}
\value{
input \code{x}, invisibly
}
\description{
Write a single data frame to SPI submission template
}
\examples{
\dontrun{
write_to_spi_sheet(camera_setup_checks_data, "my_spi_submission.xlsx")
write_to_spi_sheet(image_data, "my_spi_submission.xlsx", Surveyor = surveyor,
   `Survey Observation Photos` = file)
}
}
