% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bccamtrap-package.R
\docType{package}
\name{bccamtrap-package}
\alias{bccamtrap}
\alias{bccamtrap-package}
\title{bccamtrap: Tools for QA and validation of Camera Trap data}
\description{
Allows the easy and semi-automated checking, validation, and exploration of camera trap data collected according to BC RISC standards, and tagged using TimeLapse software.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ateucher/bccamtrap}
  \item Report bugs at \url{https://github.com/ateucher/bccamtrap/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})

Other contributors:
\itemize{
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
