% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{bccamtrapp}
\alias{bccamtrapp}
\title{Launch Shiny app}
\usage{
bccamtrapp()
}
\description{
Interactive interface for Data import, QA, and esport.
}
\examples{
\dontrun{bccamtrap_app()}

}
\concept{shiny_app}
