% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_migration_episodes}
\alias{collapse_migration_episodes}
\title{Collapse motion detection events into episodes}
\usage{
collapse_migration_episodes(df, migration_map)
}
\arguments{
\item{df}{a data frame of image data, read in with \code{read_image_data()}.}

\item{migration_map}{A small data.frame defining the upstream
direction(s) of travel for each sample station. Valid upstream values are:
\code{"L to R"}, \code{"R to L"}, \code{"T to B"}, \code{"B to T"}.
Columns must include:
\itemize{
\item \code{sample_station_label}
\item \code{upstream}
}}
}
\value{
a data frame with motion detection events collapsed into 1 row per
episode, and summary information about each episode (e.g. duration, number
of triggers, total count).
}
\description{
Collapse motion detection events into episodes
}
