% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa-deployment-images.R
\name{merge_deployments_images}
\alias{merge_deployments_images}
\title{Merge image data with deployment data}
\usage{
merge_deployments_images(deployments, image_data, as_sf = TRUE)
}
\arguments{
\item{deployments}{deployments data, as created by \code{\link[=make_deployments]{make_deployments()}}}

\item{image_data}{data.frame of class \code{image_data}, as read in by \code{\link[=read_image_data]{read_image_data()}}.}

\item{as_sf}{should the data be returned as an \code{sf} object of the station locations? Default \code{TRUE}}
}
\value{
\code{data.frame} of class \code{image_data}, with \code{deployment} columns attached
}
\description{
Attach deployment metadata with image data.
}
