% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-data.R
\name{read_image_data}
\alias{read_image_data}
\title{Read image data from a collection of csvs from TimeLapse}
\usage{
read_image_data(path, pattern, recursive = FALSE)
}
\arguments{
\item{path}{path to directory of image files.}

\item{pattern}{an optional regular expression. Only file names which match
the regular expression will read. Default \code{FALSE}.}

\item{recursive}{should files found within subfolders of \code{path} also be read?}
}
\value{
a \code{data.frame} of Timelapse image data from the files found in \code{path}
}
\description{
Read image data from a collection of csvs from TimeLapse
}
