% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_image_data}
\alias{write_image_data}
\title{Write image data to a csv or SPI submission template}
\usage{
write_image_data(
  image_data,
  file,
  ...,
  na = "",
  template = system.file("GENERIC_wildlife_camera_template_v2021.xlsm", package =
    "bccamtrap")
)
}
\arguments{
\item{image_data}{image data object.}

\item{file}{path to the output file}

\item{...}{extra columns in \code{image_data} to write out. Must be paired column
names in the form \code{"Destination Column" = image_data$data_column}. See examples.}

\item{na}{How should missing values be written. Default empty (\code{""})}

\item{template}{SPI submission template to use. Default is included in the
package, accessed by
\code{system.file("GENERIC_wildlife_camera_template_v2021.xlsm", package = "bccamtrap")}}
}
\value{
input \code{image_data}, invisibly
}
\description{
if output \code{file} specified is a csv file, this will output all of the image
data to a single csv file. If it is a \verb{xls*} file, it will populate the
\verb{Sequence Image Data} tab in a new SPI submission sheet, using a template.
}
\examples{
\dontrun{
write_image_data(image_data, "my_spi_submission.xlsx")
write_image_data(image_data, "my_spi_submission.xlsx", "Surveyor" = image_data$surveyor)
}
}
